Added Alias Feature
You can add aliases for when you want to open a program. This will save a lot of time, considering you don't have to type out the whole directory url of the file you want to execute. Finding the directory will be hard, but using "newali" after called "genalikey" will make your life easier!

Next, we'll take an array of different processes, so that with 1 alias, we can call a multitude of processes at once!